requirements.txt

Django==5.0.7
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.0
redis==5.0.1
pillow==10.4.0
qrcode[pil]==7.4.2
cryptography==46.0.5
python-decouple==3.8

